requirements.txt
discord.py-self==2.0.0
python-telegram-bot==21.10
python-dotenv==1.0.1